configs: use `systemd-networkd-wait-online --any` by default
This way, systemd-networkd-wait-online won't hang on boot for users
who only have _some_ of the configured interfaces online / in use
(which is the case for everyone who sets up Wi-Fi, since kvmd-bootconfig
does not remove eth0.network even if it was never intended to be used).

We deem this networking semantics typical for Pi-KVM in general; users
who need to wait for multiple interfaces to activate will have to
countermand this drop-in manually.

Fixes pikvm/pikvm#1514.